+++ src/index.d.ts
@@ -1,7 +1,6 @@
 import { compile } from './compile';
-import { parse } from './parse';
 declare const _default: {
     compile: typeof compile;
-    parse: typeof parse;
+    parse: (data: string) => string;
 };
 export = _default;