+++ src/parse.d.ts
@@ -1,1 +1,2 @@
-export declare function parse(data: string): string;
+import { Result } from './types';
+export declare function parse(data: string): Result;